group01/requirements.txt
autogen==0.5.3
autogen-agentchat==0.5.3
autogen-core==0.5.3
autogen-ext==0.5.3
pydantic==2.11.3
JinJa2==3.1.6
loguru==0.7.3
dynaconf==3.2.10
streamlit==1.46.0
pyyaml==6.0.1
pillow>=11.0.0

